reorganization to better support analysis
