+++ degas/products.py
@@ -1,0 +1,1 @@
+# This file will hold code to make moment maps, peak intensity maps, etc.
