+++ scripts/README
@@ -1,0 +1,1 @@
+This is where we will put random analysis scripts.
